+++ src/commands/containers/attachShellContainer.ts
@@ -8,6 +8,7 @@ import { IActionContext } from 'vscode-azureextensionui';
 import { ext } from '../../extensionVariables';
 import { ContainerTreeItem } from '../../tree/containers/ContainerTreeItem';
 import { getDockerOSType } from '../../utils/osUtils';
+import { selectAttachCommand } from '../selectCommandTemplate';
 
 export async function attachShellContainer(context: IActionContext, node?: ContainerTreeItem): Promise<void> {
     if (!node) {
@@ -29,7 +30,15 @@ export async function attachShellContainer(context: IActionContext, node?: Conta
     }
     context.telemetry.properties.shellCommand = shellCommand;
 
+    const terminalCommand = await selectAttachCommand(
+        context,
+        node.containerName,
+        node.fullTag,
+        node.containerId,
+        shellCommand
+    );
+
     const terminal = ext.terminalProvider.createTerminal(`Shell: ${node.containerName}`);
-    terminal.sendText(`docker exec -it ${node.containerId} ${shellCommand}`);
+    terminal.sendText(terminalCommand);
     terminal.show();
 }