+++ src/configureWorkspace/configureGo.ts
@@ -36,7 +36,7 @@ ${exposeStatements}
 }
 
 function genDockerCompose(serviceNameAndRelativePath: string, platform: string, os: string | undefined, ports: number[]): string {
-    return `version: '2.1'
+    return `version: '3.4'
 
 services:
   ${serviceNameAndRelativePath}:
@@ -46,7 +46,7 @@ ${getComposePorts(ports)}`;
 }
 
 function genDockerComposeDebug(serviceNameAndRelativePath: string, platform: string, os: string | undefined, ports: number[], { fullCommand: cmd }: Partial<PackageInfo>): string {
-    return `version: '2.1'
+    return `version: '3.4'
 
 services:
   ${serviceNameAndRelativePath}: