+++ src/configureWorkspace/configurePython.ts
@@ -10,8 +10,7 @@ import { TelemetryProperties } from 'vscode-azureextensionui';
 import { DockerDebugScaffoldContext } from '../debugging/DebugHelper';
 import { dockerDebugScaffoldingProvider, PythonScaffoldingOptions } from '../debugging/DockerDebugScaffoldingProvider';
 import { ext } from "../extensionVariables";
-import { PythonExtensionHelper } from '../tasks/python/PythonExtensionHelper';
-import { getPythonProjectType, inferPythonArgs, PythonDefaultDebugPort, PythonDefaultPorts, PythonFileExtension, PythonFileTarget, PythonModuleTarget, PythonProjectType, PythonTarget } from "../utils/pythonUtils";
+import { getPythonProjectType, PythonDefaultPorts, PythonFileExtension, PythonFileTarget, PythonModuleTarget, PythonProjectType, PythonTarget } from "../utils/pythonUtils";
 import { getComposePorts, getExposeStatements } from './configure';
 import { ConfigureTelemetryProperties, genCommonDockerIgnoreFile, quickPickGenerateComposeFiles } from './configUtils';
 import { ScaffolderContext, ScaffoldFile } from './scaffolding';
@@ -22,15 +21,21 @@ interface LaunchFilePrompt {
 }
 
 const defaultLaunchFile: Map<PythonProjectType, LaunchFilePrompt> = new Map<PythonProjectType, LaunchFilePrompt>([
-    ['django', { prompt: 'Enter the relative path to the application (e.g. manage.py)', defaultFile: 'manage.py' }],
-    ['flask', { prompt: 'Enter the relative path to the application (e.g. \'app.py\' or \'app\')', defaultFile: 'app.py' }],
-    ['general', { prompt: 'Enter the relative path to the application (e.g. \'app.py\' or \'app\')', defaultFile: 'app.py' }],
+    ['django', { prompt: 'Enter the relative path to the app’s entry point (e.g. manage.py or subfolder_name/manage.py)', defaultFile: 'manage.py' }],
+    ['flask', { prompt: 'Enter the relative path to the app’s entry point (e.g. app.py or subfolder_name/app.py)', defaultFile: 'app.py' }],
+    ['general', { prompt: 'Enter the relative path to the app’s entry point (e.g. app.py or subfolder_name/app.py)', defaultFile: 'app.py' }],
 ]);
 
 const pythonDockerfile = `# For more information, please refer to https://aka.ms/vscode-docker-python
-FROM python
+FROM python:3.8
 
 $expose_statements$
+
+# Keeps Python from generating .pyc files in the container
+ENV PYTHONDONTWRITEBYTECODE 1
+
+# Turns off buffering for easier container logging
+ENV PYTHONUNBUFFERED 1
 
 # Install pip requirements
 ADD requirements.txt .
@@ -52,25 +57,13 @@ services:
       dockerfile: Dockerfile
 $ports$`;
 
-const dockerComposeDebugfile = `version: '3.4'
+const djangoRequirements = `django==3.0.3
+gunicorn==20.0.4`;
 
-services:
-  $service_name$:
-    image: $service_name$
-    build:
-      context: .
-      dockerfile: Dockerfile
-$dbg_volume$
-    entrypoint: $entrypoint$
-$ports$`;
+const flaskRequirements = `flask==1.1.1
+gunicorn==20.0.4`;
 
-const djangoRequirements = `django
-gunicorn`;
-
-const flaskRequirements = `flask
-gunicorn`;
-
-function genDockerFile(serviceName: string, target: PythonTarget, projectType: PythonProjectType, ports: number[]): string {
+async function genDockerFile(serviceName: string, target: PythonTarget, projectType: PythonProjectType, ports: number[], outputFolder: string): Promise<string> {
     const exposeStatements = getExposeStatements(ports);
     let command = '';
 
@@ -81,8 +74,7 @@ function genDockerFile(serviceName: string, target: PythonTarget, projectType: P
             command = `CMD ["python", "-m", "${(target as PythonModuleTarget).module}"]`;
         }
     } else if (projectType === 'django') {
-        // For Django apps, there **usually** exists a "wsgi" module, so our best guess is to use the folder name.
-        command = `CMD ["gunicorn", "--bind", "0.0.0.0:${ports ? ports[0] : PythonDefaultPorts[projectType]}", "${serviceName}.wsgi"]`;
+        command = await inferDjangoCommand(ports, serviceName, outputFolder);
     } else if (projectType === 'flask') {
         // For Flask apps, our guess is to assume there is a callable "app" object in the file/module that the user provided.
         command = `CMD ["gunicorn", "--bind", "0.0.0.0:${ports ? ports[0] : PythonDefaultPorts[projectType]}", "${inferPythonWsgiModule(target)}:app"]`;
@@ -102,40 +94,15 @@ function genDockerCompose(serviceName: string, ports: number[]): string {
         .replace(/\$ports\$/g, getComposePorts(ports));
 }
 
-async function genDockerComposeDebug(serviceName: string, projectType: PythonProjectType, ports: number[], target: PythonTarget): Promise<string> {
-    const defaultDebugOptions : PythonExtensionHelper.DebugLaunchOptions = {
-        host: '0.0.0.0',
-        port: PythonDefaultDebugPort,
-        wait: true
-    };
-
-    const args = inferPythonArgs(projectType, ports);
-    const launcherCommand = PythonExtensionHelper.getRemotePtvsdCommand(target, args, defaultDebugOptions);
-    const entrypoint = 'python '.concat(launcherCommand);
-    const launcherFolder = await PythonExtensionHelper.getLauncherFolderPath();
-
-    let dbgVolume : string = '';
-
-    if (launcherFolder) {
-        dbgVolume = `    volumes:\n      - ${launcherFolder}:/pydbg`;
-    }
-
-    return dockerComposeDebugfile
-        .replace(/\$service_name\$/g, serviceName)
-        .replace(/\$dbg_volume\$/g, dbgVolume)
-        .replace(/\$entrypoint\$/g, entrypoint)
-        .replace(/\$ports\$/g, getComposePorts(ports, PythonDefaultDebugPort));
-}
-
 function genRequirementsFile(projectType: PythonProjectType): string {
-    let contents = '# Add requirements when needed'
+    let contents = '# To ensure app dependencies are ported from your virtual environment/host machine into your container, run \'pip freeze > requirements.txt\' in the terminal to overwrite this file';
 
     switch (projectType) {
         case 'django':
-            contents = djangoRequirements;
+            contents = contents.concat('\n', djangoRequirements);
             break;
         case 'flask':
-            contents = flaskRequirements
+            contents = contents.concat('\n', flaskRequirements);
             break;
         default:
     }
@@ -143,14 +110,13 @@ function genRequirementsFile(projectType: PythonProjectType): string {
     return contents;
 }
 
-async function initializeForDebugging(context: ScaffolderContext, dockerfile: string, ports: number[], generateComposeFiles: boolean,
+async function initializeForDebugging(context: ScaffolderContext, dockerfile: string, ports: number[],
                                       target: PythonTarget, projectType: PythonProjectType): Promise<void> {
     const scaffoldContext: DockerDebugScaffoldContext = {
         folder: context.folder,
         platform: 'python',
         actionContext: context,
         dockerfile: dockerfile,
-        generateComposeTask: generateComposeFiles,
         ports: ports
     }
 
@@ -176,6 +142,21 @@ function inferPythonWsgiModule(target: PythonTarget): string {
     return wsgiModule.replace(/\//g, '.');
 }
 
+async function inferDjangoCommand(ports: number[], serviceName: string, outputFolder: string) : Promise<string> {
+    // For Django apps, there **usually** exists a "wsgi" module in a sub-folder named the same as the project folder.
+    // So we check if that path exists, then use it. Else, we output the comment below instructing the user to enter
+    // the correct python path to the wsgi module.
+
+    const wsgiPath = path.join(outputFolder, path.join(serviceName, "wsgi.py"));
+    const command = `CMD ["gunicorn", "--bind", "0.0.0.0:${ports ? ports[0] : PythonDefaultPorts.get('django')}", "$wsgi_path$"]`;
+
+    if (await fse.pathExists(wsgiPath)) {
+        return command.replace(/\$wsgi_path\$/g, `${serviceName}.wsgi`);
+    } else {
+        return `# File wsgi.py was not found in subfolder:${serviceName}. Please enter the Python path to wsgi file.`
+            .concat('\n', command.replace(/\$wsgi_path\$/g, 'pythonPath.to.wsgi'));
+    }
+}
 
 export async function promptForLaunchFile(projectType?: PythonProjectType) : Promise<PythonTarget> {
     const launchFilePrompt = defaultLaunchFile.get(projectType);
@@ -187,11 +168,13 @@ export async function promptForLaunchFile(projectType?: PythonProjectType) : Pro
         validateInput: (value: string): string | undefined => { return value && value.trim().length > 0 ? undefined : 'Enter a valid Python file path/module.' }
     };
 
-    const file = await ext.ui.showInputBox(opt);
+    // Ensure to change any \ to /.
+    const file = (await ext.ui.showInputBox(opt)).replace(/\\/g, '/');
 
-    // If the input has the .py extension, then assume it is a file.
-    if (path.extname(file).toLocaleUpperCase() === PythonFileExtension.toLocaleUpperCase()) {
-        return { file: file.replace(/\\/g, '/') };
+    // If the input has the .py extension or a forward-slash, then treat it as a file/directory (i.e. execute without the -m flag).
+    if (path.extname(file).toLocaleUpperCase() === PythonFileExtension.toLocaleUpperCase() ||
+        file.indexOf('/') > 0) {
+        return { file: file };
     } else {
         return { module: file};
     }
@@ -205,6 +188,7 @@ export async function scaffoldPython(context: ScaffolderContext): Promise<Scaffo
 
     const generateComposeFiles = await context.captureStep('compose', quickPickGenerateComposeFiles)();
     const projectType = getPythonProjectType(context.platform);
+    const launchFile = await context.captureStep('pythonFile', promptForLaunchFile)(projectType);
 
     const defaultPort = PythonDefaultPorts.get(projectType);
     let ports = [];
@@ -213,12 +197,12 @@ export async function scaffoldPython(context: ScaffolderContext): Promise<Scaffo
         ports = await context.promptForPorts([ defaultPort ]);
     }
 
-    const launchFile = await context.captureStep('pythonFile', promptForLaunchFile)(projectType);
-    const dockerFileContents = genDockerFile(serviceName, launchFile, projectType, ports);
+    const dockerFileContents = await genDockerFile(serviceName, launchFile, projectType, ports, outputFolder);
+    const dockerIgnoreContents = '**/__pycache__\n'.concat(genCommonDockerIgnoreFile(context.platform));
 
     const files: ScaffoldFile[] = [
         { fileName: 'Dockerfile', contents: dockerFileContents, open: true },
-        { fileName: '.dockerignore', contents: genCommonDockerIgnoreFile(context.platform) }
+        { fileName: '.dockerignore', contents: dockerIgnoreContents }
     ];
 
     const requirementsFileExists = await fse.pathExists(path.join(outputFolder, 'requirements.txt'));
@@ -231,11 +215,8 @@ export async function scaffoldPython(context: ScaffolderContext): Promise<Scaffo
         properties.orchestration = 'docker-compose';
 
         const dockerComposeFile = genDockerCompose(serviceName, ports);
-        const dockerComposeDebugFile = await genDockerComposeDebug(serviceName, projectType, ports, launchFile);
 
-        files.push(
-            { fileName: 'docker-compose.yml', contents: dockerComposeFile },
-            { fileName: 'docker-compose.debug.yml', contents: dockerComposeDebugFile });
+        files.push({ fileName: 'docker-compose.yml', contents: dockerComposeFile });
     }
 
     files.forEach(file => {
@@ -247,7 +228,7 @@ export async function scaffoldPython(context: ScaffolderContext): Promise<Scaffo
     });
 
     if (context.initializeForDebugging) {
-        await initializeForDebugging(context, path.join(outputFolder, 'Dockerfile'), ports, generateComposeFiles, launchFile, projectType);
+        await initializeForDebugging(context, path.join(outputFolder, 'Dockerfile'), ports, launchFile, projectType);
     }
 
     return files;