+++ src/debugging/netcore/NetCoreDebugHelper.ts
@@ -6,13 +6,16 @@
 import * as fse from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
-import { DebugConfiguration } from 'vscode';
+import { DebugConfiguration, MessageItem, window } from 'vscode';
+import { DialogResponses, IActionContext, UserCancelledError } from 'vscode-azureextensionui';
 import { ext } from '../../extensionVariables';
 import { NetCoreTaskHelper, NetCoreTaskOptions } from '../../tasks/netcore/NetCoreTaskHelper';
+import { ContainerTreeItem } from '../../tree/containers/ContainerTreeItem';
 import { pathNormalize } from '../../utils/pathNormalize';
 import { PlatformOS } from '../../utils/platform';
 import { unresolveWorkspaceFolder } from '../../utils/resolveVariables';
 import { ChildProcessProvider } from '../coreclr/ChildProcessProvider';
+import CliDockerClient, { DockerExecOptions } from '../coreclr/CliDockerClient';
 import { CommandLineDotNetClient } from '../coreclr/CommandLineDotNetClient';
 import { LocalFileSystemProvider } from '../coreclr/fsProvider';
 import { AspNetCoreSslManager, LocalAspNetCoreSslManager } from '../coreclr/LocalAspNetCoreSslManager';
@@ -22,10 +25,11 @@ import { DefaultOutputManager } from '../coreclr/outputManager';
 import { OSTempFileProvider } from '../coreclr/tempFileProvider';
 import { RemoteVsDbgClient, VsDbgClient } from '../coreclr/vsdbgClient';
 import { DebugHelper, DockerDebugContext, DockerDebugScaffoldContext, inferContainerName, ResolvedDebugConfiguration, resolveDockerServerReadyAction } from '../DebugHelper';
-import { DockerDebugConfiguration } from '../DockerDebugConfigurationProvider';
+import { DockerAttachConfiguration, DockerDebugConfiguration } from '../DockerDebugConfigurationProvider';
 
 export interface NetCoreDebugOptions extends NetCoreTaskOptions {
     appOutput?: string;
+    debuggerPath?: string;
 }
 
 export interface NetCoreDockerDebugConfiguration extends DebugConfiguration {
@@ -99,6 +103,19 @@ export class NetCoreDebugHelper implements DebugHelper {
     }
 
     public async resolveDebugConfiguration(context: DockerDebugContext, debugConfiguration: DockerDebugConfiguration): Promise<ResolvedDebugConfiguration | undefined> {
+        switch (debugConfiguration.request) {
+            case 'launch':
+                return this.resolveLauchDebugConfiguration(context, debugConfiguration);
+                break;
+            case 'attach':
+                return this.resolveAttachDebugConfiguration(context, debugConfiguration);
+                break;
+            default:
+                throw Error(`Unknown request ${debugConfiguration.request} specified in the debug config.`);
+        }
+    }
+
+    public async resolveLauchDebugConfiguration(context: DockerDebugContext, debugConfiguration: DockerDebugConfiguration): Promise<ResolvedDebugConfiguration | undefined> {
         debugConfiguration.netCore = debugConfiguration.netCore || {};
         debugConfiguration.netCore.appProject = await NetCoreTaskHelper.inferAppProject(context.folder, debugConfiguration.netCore); // This method internally checks the user-defined input first
 
@@ -167,6 +184,37 @@ export class NetCoreDebugHelper implements DebugHelper {
         };
     }
 
+    public async resolveAttachDebugConfiguration(context: DockerDebugContext, debugConfiguration: DockerAttachConfiguration): Promise<ResolvedDebugConfiguration | undefined> {
+        // TODO: Validate the target container OS and fail debugging
+        // Get Container Name if missing
+        const containerName: string = debugConfiguration.containerName ?? await this.getContainerNameToAttach();
+
+        // If debugger path is not specified, install the debugger
+        const debuggerPath: string = debugConfiguration.netCore?.debuggerPath ?? await this.installDebuggerInContainer(containerName);
+
+        return {
+            ...debugConfiguration, // Gets things like name, preLaunchTask, serverReadyAction, etc.
+            type: 'coreclr',
+            request: 'attach',
+            'justMyCode': false,
+            // if processId is specified in the debugConfiguration, then it will take precedences
+            // and processName will be ignored.
+            processName: debugConfiguration.processName || 'dotnet',
+            pipeTransport: {
+                pipeProgram: 'docker',
+                pipeArgs: ['exec', '-i', containerName],
+                // eslint-disable-next-line no-template-curly-in-string
+                pipeCwd: '${workspaceFolder}',
+                debuggerPath: debuggerPath,
+                quoteArgs: false,
+            },
+            sourceFileMap: debugConfiguration.sourceFileMap || {
+                // eslint-disable-next-line no-template-curly-in-string
+                '/src': '${workspaceFolder}'
+            }
+        };
+    }
+
     public static getHostDebuggerPathBase(): string {
         return path.join(os.homedir(), '.vsdbg');
     }
@@ -227,6 +275,49 @@ export class NetCoreDebugHelper implements DebugHelper {
             path.posix.join('/app', relativePath);
 
         return pathNormalize(result, platformOS);
+    }
+
+    private async installDebuggerInContainer(containerName: string): Promise<string> {
+        const yesItem: MessageItem = DialogResponses.yes;
+        const install = (yesItem === await window.showInformationMessage('Attaching to container requires .NET Core debugger in the container. Do you want to install debugger in the container?', ...[DialogResponses.yes, DialogResponses.no]));
+        if (!install) {
+            throw new UserCancelledError("User didn't grand permission to install .NET Core debugger.");
+        }
+
+        const debuggerPath: string = '/remote_debugger';
+        // Windows require double quotes and Mac and Linux require single quote.
+        const osProvider = new LocalOSProvider();
+        const installDebugger: string = osProvider.os === 'Windows' ?
+            `/bin/sh -c "ID=default; if [ -e /etc/os-release ]; then . /etc/os-release; fi; echo $ID; if [ $ID == alpine ]; then apk --no-cache add curl && curl -sSL https://aka.ms/getvsdbgsh | /bin/sh /dev/stdin -v latest -l ${debuggerPath} ; else apt-get update && apt-get install unzip && curl -sSL https://aka.ms/getvsdbgsh | /bin/sh /dev/stdin -v latest -l ${debuggerPath}; fi"`
+            : `/bin/sh -c 'ID=default; if [ -e /etc/os-release ]; then . /etc/os-release; fi; echo $ID; if [ $ID == alpine ]; then apk --no-cache add curl && curl -sSL https://aka.ms/getvsdbgsh | /bin/sh /dev/stdin -v latest -l ${debuggerPath} ; else apt-get update && apt-get install unzip && curl -sSL https://aka.ms/getvsdbgsh | /bin/sh /dev/stdin -v latest -l ${debuggerPath}; fi'`
+
+        const outputManager = new DefaultOutputManager(ext.outputChannel);
+        const dockerClient = new CliDockerClient(new ChildProcessProvider());
+
+        await outputManager.performOperation(
+            'Installing the latest .NET Core debugger...',
+            async (output) => {
+                const installProgress = (content: string) => {
+                    output.appendLine(content);
+                };
+                const dockerExecOptions : DockerExecOptions = { interactive: true, progress: installProgress };
+
+                await dockerClient.exec(containerName, installDebugger, dockerExecOptions);
+            },
+            'Debugger installed',
+            'Unable to install the .NET Core debugger.'
+        );
+
+        return `${debuggerPath}/vsdbg`;
+    }
+
+    private async getContainerNameToAttach(): Promise<string> {
+        const context: IActionContext = { telemetry: { properties: {}, measurements: {} }, errorHandling: { issueProperties: {} } };
+        const containerItem: ContainerTreeItem = await ext.containersTree.showTreeItemPicker(ContainerTreeItem.runningContainerRegExp, {
+            ...context,
+            noItemFoundErrorMessage: 'No running containers are available to attach'
+        });
+        return containerItem.containerName;
     }
 }